% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pipe_source.R
\name{get_pipe_source}
\alias{get_pipe_source}
\title{nm_mean <- mean %>% set_namer}
\usage{
get_pipe_source(x)
}
\arguments{
\item{x}{magrittr}
}
\description{
"price" %>% name_it( nm_mean )
Get source argument from a chain
}
\details{
Magrittr pipe function



 `lazyeval::expr_text` does not work for pipes ...
}
\examples{

 f <- . \%>\% mean( na.rm = TRUE )
 nm_mean <-  mean \%>\% set_namer()

 get_pipe_source(f)

}
\seealso{
- `lazyeval::expr_text`
}
