% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_pipe.R
\name{inside_pipe}
\alias{inside_pipe}
\title{inside_pipe}
\usage{
inside_pipe(...)
}
\arguments{
\item{...}{arguments}
}
\value{
logical
}
\description{
Detects if the function is being used in a magrittr pipe
}
\details{
Examines parent.frames to see if the function is being used in a magrittr
pipe. This is useful when the presence of a pipe effects what you need to
do.
}
\examples{

"foo" \%>\% ( function(x) inside_pipe() )
f <- function(x) inside_pipe(x)

"foo" \%>\% f

}
