% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\docType{methods}
\name{\%>\%<-}
\alias{\%>\%<-}
\alias{\%>\%<-,ANY-method}
\alias{'\%>\%<-',}
\alias{ANY}
\title{Pipe Replacement}
\usage{
\%>\%(...) <- value

\S4method{\%>\%}{ANY}(...) <- value
}
\arguments{
\item{value}{any; value for the LHS}

\item{...}{additional arguments}
}
\description{
Pipe Replacement
}
\examples{
  ll <- list( a=1, b=2 )
  ll \%>\% a <- 3
  ll$a

}
