% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-dots.R
\name{..}
\alias{..}
\title{two_dots}
\usage{
..(.ignored, x, y = parent.frame(7))
}
\arguments{
\item{x}{name; unquote name}

\item{y}{envir}

\item{.ignore}{unused useful for pipes}
}
\description{
Find value in parent environment
}
\examples{

 ll <- list( a=1:3, b=7)
 a = "ok"
 ll \%>\% a
 ll \%>\% ..(a)

}
\references{
\url{http://stackoverflow.com/a/21664128} applies the same principal to
  data.tables
}
