% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap_function.R
\name{wrap_function}
\alias{wrap_function}
\title{Wrap an expression in a function}
\usage{
wrap_function(body, pipe, env)
}
\arguments{
\item{body}{}

\item{pipe}{character; the pipe used}

\item{env}{environment;

This is a monkey patch of magrittr's wrap_function that extends the
functionality to perform non-standard-evaluation and introspect the RHS of
pipe within the LHS.}
}
\description{
Wrap an expression in a function
}
\examples{
ll <- list( a=1, b=2 )
  ll \%>\% a
  ll \%>\% b
}
\author{
magrittr authors + Chris Brown
}

