% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replacement.R
\docType{methods}
\name{\%>\%<-}
\alias{\%>\%<-}
\alias{\%>\%<-,ANY-method}
\title{Pipe Replacement}
\usage{
\%>\%(...) <- value

\S4method{\%>\%}{ANY}(...) <- value
}
\arguments{
\item{value}{any; value to set the LHS to}

\item{...}{additional arguments}

\item{value}{ANY}
}
\description{
Pipe Replacement
}
\examples{
ll <- list( a=1, b=2 )
  ll \%>\% a <- 3
  ll$a
}

